connectivity: the 64-bit port missed this file, and it writes through the truncated pointers
wmt_ic_soc.c casts kernel pointers through UINT32 before storing them in
au4CtrlData, which is SIZE_T[] — 64-bit here. The consumer casts them straight
back and dereferences them. wmt_ctrl.c, wmt_ctrl_get_patch():

    PPUINT8 ppBuf = (PPUINT8) pWmtCtrlData->au4CtrlData[2];
    PUINT32 pSize = (PUINT32) pWmtCtrlData->au4CtrlData[3];
    pFullPatchName = (PUINT8) pWmtCtrlData->au4CtrlData[1];
    ...
    *ppBuf = (PUINT8)(pPatch)->data;     <- eight bytes
    *pSize = (pPatch)->size;             <- four bytes

and the producer, mtk_wcn_soc_patch_dwn():

    ctrlData.au4CtrlData[1] = (UINT32) &gFullPatchName;
    ctrlData.au4CtrlData[2] = (UINT32) &pbuf;        <- kernel stack address
    ctrlData.au4CtrlData[3] = (UINT32) &patchSize;   <- kernel stack address

&pbuf and &patchSize are stack addresses; the cast discards the top 32 bits, so
those two writes land at a low-4GiB address — the userspace range — from kernel
context, with PAN clear. Once per CONSYS func-on, which is once per boot, on
every build this port has ever made.

THIS IS THE LIVE PATH, not dead code. From the boot that carries this fix:

  [133.561698] [WMT-IC][I]mtk_wcn_soc_patch_dwn:[Patch]BuiltTime = 20180615...
  [133.623294] [WMT-IC][W]mtk_wcn_soc_patch_dwn:wmt_core: patch dwn:0 frag(47, 444) ok

(The other four casts in this file are under CFG_WMT_CRYSTAL_TIMING_SET, which
is 0, and are fixed anyway.)

HOW IT WAS FOUND, because the method generalises: DEBUG_PAGEALLOC caught nothing
through a full teardown, which argues against a page use-after-free and left the
issue's own reading of the fault address — "a 64-bit pointer read out of a
location that holds a 32-bit address packed next to an unrelated 32-bit word,
the failure mode you would expect from a driver ported from a 32-bit 3.18 tree"
— as the thing to test STATICALLY rather than wait for.

The tell is that the port was done unevenly. wmt_ic_6628.c performs the
identical operations and already says (size_t) on every one of them; this file
was not updated with it, and this file is the one the SoC builds (Makefile:
mtk_stp_wmt-objs includes core/wmt_ic_soc.o; 6620 and 6628 are not built at
all). p2p_func.c line 3854 says (ULONG) for an expression that line 3086 wrote
with (UINT_32) — same file, same idiom, one fixed and one missed. Both are here.

VERIFIED, not just compiled: rebuilt, installed over the shadowing copies in
/lib/modules/.../updates/, rebooted. Wi-Fi associates, the patch download
completes, no faults, no failed units.

WHAT THIS DOES NOT CLAIM. That it is #72's root cause is unproven — the
observed fault address does not obviously come from these writes, and what they
corrupt is whatever happens to be mapped at a truncated stack address. It is a
definite memory-corrupting bug on the exact path where seven of eight #72
faults clustered, fixed. Whether #72 stops is now an observation, and the soak
is running.

+++ modules/connectivity/common_main/core/wmt_ic_soc.c
@@ -1933,9 +1933,9 @@ static INT32 mtk_wcn_soc_crystal_triming_set(VOID)
 
 	bIsNvramExist = MTK_WCN_BOOL_FALSE;
 	 /**/ ctrlData.ctrlId = WMT_CTRL_CRYSTAL_TRIMING_GET;
-	ctrlData.au4CtrlData[0] = (UINT32) "/data/nvram/APCFG/APRDEB/WIFI";
-	ctrlData.au4CtrlData[1] = (UINT32) &pbuf;
-	ctrlData.au4CtrlData[2] = (UINT32) &bufLen;
+	ctrlData.au4CtrlData[0] = (size_t) "/data/nvram/APCFG/APRDEB/WIFI";
+	ctrlData.au4CtrlData[1] = (size_t) &pbuf;
+	ctrlData.au4CtrlData[2] = (size_t) &bufLen;
 
 	iRet = wmt_ctrl(&ctrlData);
 	if (0 != iRet) {
@@ -1965,7 +1965,7 @@ static INT32 mtk_wcn_soc_crystal_triming_set(VOID)
 				     bIsCrysTrimEnabled);
 		}
 		ctrlData.ctrlId = WMT_CTRL_CRYSTAL_TRIMING_PUT;
-		ctrlData.au4CtrlData[0] = (UINT32) "/data/nvram/APCFG/APRDEB/WIFI";
+		ctrlData.au4CtrlData[0] = (size_t) "/data/nvram/APCFG/APRDEB/WIFI";
 		iRet = wmt_ctrl(&ctrlData);
 		if (0 != iRet) {
 			WMT_ERR_FUNC("0x%x: WMT_CTRL_CRYSTAL_TRIMING_PUT fail:%d\n", wmt_ic_ops_soc.icId, iRet);
@@ -2374,21 +2374,43 @@ static INT32 mtk_wcn_soc_patch_dwn(VOID)
 	if (0 == iRet) {
 		/* patch with correct Hw Ver Major Num found */
 		ctrlData.ctrlId = WMT_CTRL_GET_PATCH_NAME;
-		ctrlData.au4CtrlData[0] = (UINT32) &gFullPatchName;
+/*
+		 * (size_t), NOT (UINT32) -- these are POINTERS.
+		 *
+		 * au4CtrlData is SIZE_T[] (wmt_ctrl.h), i.e. 64-bit here, and wmt_ctrl_get_patch()
+		 * casts these entries straight back to pointers and WRITES THROUGH THEM:
+		 *
+		 *     PPUINT8 ppBuf = (PPUINT8) pWmtCtrlData->au4CtrlData[2];
+		 *     PUINT32 pSize = (PUINT32) pWmtCtrlData->au4CtrlData[3];
+		 *     ...
+		 *     *ppBuf = (PUINT8)(pPatch)->data;
+		 *     *pSize = (pPatch)->size;
+		 *
+		 * &pbuf and &patchSize are kernel STACK addresses (0xffff8000_xxxxxxxx on
+		 * arm64). Casting through UINT32 discards the top half, so the consumer writes
+		 * eight bytes and four bytes through a pointer that is now in the low 4 GiB --
+		 * the userspace range -- on every CONSYS func-on.
+		 *
+		 * This is a 32-bit-era cast that the 64-bit port missed. The identical
+		 * operations in wmt_ic_6628.c already say (size_t); this file was not updated
+		 * with it, and this file is the one this SoC builds (Makefile: mtk_stp_wmt-objs
+		 * includes core/wmt_ic_soc.o, and 6620/6628 are not built at all).
+		 */
+		ctrlData.au4CtrlData[0] = (size_t) &gFullPatchName;
 		iRet = wmt_ctrl(&ctrlData);
 
 		WMT_INFO_FUNC("valid patch found: (%s)\n", gFullPatchName);
 		/* <2.2> read patch content */
 		ctrlData.ctrlId = WMT_CTRL_GET_PATCH;
 		ctrlData.au4CtrlData[0] = (UINT32) NULL;
-		ctrlData.au4CtrlData[1] = (UINT32) &gFullPatchName;
+		ctrlData.au4CtrlData[1] = (size_t) &gFullPatchName;
 
 	} else {
 		iRet -= 1;
 		return iRet;
 	}
-	ctrlData.au4CtrlData[2] = (UINT32) &pbuf;
-	ctrlData.au4CtrlData[3] = (UINT32) &patchSize;
+	ctrlData.au4CtrlData[2] = (size_t) &pbuf;
+	ctrlData.au4CtrlData[3] = (size_t) &patchSize;
 	iRet = wmt_ctrl(&ctrlData);
 	if (iRet) {
 		WMT_ERR_FUNC("wmt_core: WMT_CTRL_GET_PATCH fail:%d\n", iRet);
