+++ modules/connectivity/common_main/linux/wmt_dbg.c
@@ -734,7 +734,8 @@ static INT32 wmt_internal_loopback(INT32 count, INT32 max)
 		pSignal = &pOp->signal;
 		pOp->op.opId = WMT_OPID_LPBK;
 		pOp->op.au4OpData[0] = lpbk_buffer.payload_length;	/* packet length */
-		pOp->op.au4OpData[1] = (UINT32) &gLpbkBuf[0];
+		/* (size_t): au4OpData is SIZE_T[] and this is a pointer. */
+		pOp->op.au4OpData[1] = (size_t) &gLpbkBuf[0];
 		pSignal->timeoutValue = MAX_EACH_WMT_CMD;
 		WMT_INFO_FUNC("OPID(%d) type(%d) start\n", pOp->op.opId, pOp->op.au4OpData[0]);
 		if (DISABLE_PSM_MONITOR()) {
